\name{.normal.plot}
\alias{.normal.plot}
\title{Plot Simualted Quantities of Interest for GAM Zelig Models...}
\usage{\method{plot}{sim.logit.gam}(x, ...)}
\description{Plot Simualted Quantities of Interest for GAM Zelig Models}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a simulated GEE object containing simulations of quantities of
interest}
\item{...}{ignored parameters}}
