\name{.plot.density}
\alias{.plot.density}
\title{Plot Simualted Quantities of Interest for GEE Zelig Models...}
\usage{\method{plot}{sim.probit.gee}(x, ...)}
\description{Plot Simualted Quantities of Interest for GEE Zelig Models}
\author{Matt Owen \email{mowen@iq.harvard.edu}
Plot Density Graphs for GEE Quantities of Interest}
\arguments{\item{x}{a simulated GEE object containing simulations of quantities of
interest}
\item{...}{ignored parameters}
\item{x}{a vector containing quantities of interest}
\item{main}{the main title of the plot}
\item{col}{the color of the line-plot}}
