\name{describe.logit.gam}
\alias{describe.logit.gam}
\title{Describe the \code{logit.gam} model to Zelig
This function creates the citation information for the ``logit.gam'' model.}
\usage{
  \method{describe}{logit.gam}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list of important information
}
\description{
  Describe the \code{logit.gam} model to Zelig This
  function creates the citation information for the
  ``logit.gam'' model.
}
\author{
  Skyler J. Cranmer
}

