\name{describe.poisson.gam}
\alias{describe.poisson.gam}
\title{Describe the \code{poisson.gam} model to Zelig
This function creates the citation information for the ``poisson.gam'' model.}
\usage{
  \method{describe}{poisson.gam}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list of important information
}
\description{
  Describe the \code{poisson.gam} model to Zelig This
  function creates the citation information for the
  ``poisson.gam'' model.
}
\author{
  Skyler J. Cranmer
}

