\name{describe.probit.gam}
\alias{describe.probit.gam}
\title{Describe the \code{probit.gam} model to Zelig
This function creates the citation information for the ``probit.gam'' model.}
\usage{
  \method{describe}{probit.gam}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list of important information
}
\description{
  Describe the \code{probit.gam} model to Zelig This
  function creates the citation information for the
  ``probit.gam'' model.
}
\author{
  Skyler J. Cranmer
}

