\name{gam.zelig-package}
\title{gam...}
\description{gam.zelig}
\details{\tabular{ll}{
Package: \tab gam.zelig\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-25\cr
Depends: Zelig, gam, mgcv
License: \tab GPL version 2 or newer\cr
}

Edit this description}
\alias{gam.zelig-package}
\alias{gam.zelig}
\docType{package}
\author{Skyler Cranmer}
\keyword{package}
