\name{GAMplot}
\alias{GAMplot}
\title{Plot Either Bar or Density Plot...}
\usage{GAMplot(x, main, col)}
\description{Plot Either Bar or Density Plot}
\value{NULL (invisibly)}
\arguments{\item{x}{a vector containing quantities of interest}
\item{main}{the main title of the plot}
\item{col}{the color of the line-plot}}
