\name{model.frame.GamFormula}
\alias{model.frame.GamFormula}
\title{Extract Model Frame from GAM-style Formula...}
\usage{\method{model.frame}{GamFormula}(formula, data, ...)}
\description{Extract Model Frame from GAM-style Formula}
\details{Produce a \code{model.frame} from a a GAM-style formula and a
\code{data.frame}.}
\author{Matt Owen \email{mowen@iq.harvard.edu} and Skyler J. Cranmer}
\arguments{\item{formula}{a gamma-style formula with class \code{GamFormula}}
\item{data}{a \code{data.frame} used to create the rows of the
\code{model.frame}}
\item{...}{ignored parameters}}
