\name{model.frame.GammaFormula}
\alias{model.frame.GammaFormula}
\title{Extract Model Frame from Gamma-style Formula...}
\usage{model.frame.GammaFormula(formula, data, ...)}
\description{Extract Model Frame from Gamma-style Formula}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{formula}{}
\item{data}{}
\item{...}{}}
