\name{plotdensity}
\alias{plotdensity}
\title{Plot Density Graphs for GEE Quantities of Interest...}
\usage{plotdensity(x, main, col)}
\description{Plot Density Graphs for GEE Quantities of Interest}
\arguments{\item{x}{a vector containing quantities of interest}
\item{main}{the main title of the plot}
\item{col}{the color of the line-plot}}
