\name{plotfactor}
\alias{plotfactor}
\title{Plot Density Graphs for GEE Quantities of Interest}
\usage{
  plotfactor(x, main, col)
}
\arguments{
  \item{x}{a vector containing quantities of interest}

  \item{main}{the main title of the plot}

  \item{col}{the color of the line-plot}
}
\description{
  Plot Density Graphs for GEE Quantities of Interest
}

