\name{zelig2gam.logit}
\alias{zelig2gam.logit}
\title{Interface between the Zelig Model gam...}
\usage{zelig2gam.logit(formula, ..., data)}
\description{Interface between the Zelig Model gam.logit and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
