\name{zelig2gam.poisson}
\alias{zelig2gam.poisson}
\title{Interface between the Zelig Model gam...}
\usage{zelig2gam.poisson(formula, ..., data)}
\description{Interface between the Zelig Model gam.poisson and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
