\name{zelig2normal.gam}
\alias{zelig2normal.gam}
\title{Interface between the Zelig Model normal.gam and
the Pre-existing Model-fitting Method}
\usage{
  zelig2normal.gam(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model normal.gam and the
  Pre-existing Model-fitting Method
}

