\docType{package}
\name{ZeligGAM-package}
\alias{ZeligGAM}
\alias{ZeligGAM-package}
\title{ZeligGAM}
\description{
  \tabular{ll}{ Package: \tab ZeligGAM\cr Version: \tab
  0.7\cr Date: \tab 2012-10-02\cr Depends: Zelig (>=
  4.0-11), gam, mgcv\cr License: \tab GPL version 2 or
  newer\cr }
}
\details{
  General Additive Models extend GLM's to support the
  application of smoothing functions to the predictor
  terms.
}
\author{
  Skyler Cranmer
}
\keyword{Additive}
\keyword{GAM,}
\keyword{Generalized}
\keyword{Models,}
\keyword{package,}
\keyword{Zelig}

