\name{param.poisson.gam}
\alias{param.poisson.gam}
\title{Extract Samples from a Distribution in Order to Pass Them to the qi Function...}
\description{Extract Samples from a Distribution in Order to Pass Them to the \code{qi} Function
(this is primarily a helper function for the poisson.gam model)}
\value{a list specifying link, link-inverse, random samples, and ancillary parameters}
\arguments{\item{obj}{a zelig object}
\item{num}{an integer specifying the number of simulations to compute}
\item{...}{additional parameters}}
