\name{plot.sim.poisson.gam}
\alias{plot.sim.poisson.gam}
\title{Plot Simualted Quantities of Interest for GAM Zelig Models...}
\usage{\method{plot}{sim.poisson.gam}(x, ...)}
\description{Plot Simualted Quantities of Interest for GAM Zelig Models}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a simulated GEE object containing simulations of quantities of
interest}
\item{...}{ignored parameters}}
