\name{zelig2logit.gam}
\alias{zelig2logit.gam}
\title{Interface between the Zelig Model logit.gam and
the Pre-existing Model-fitting Method}
\usage{
  zelig2logit.gam(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additional parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model logit.gam and the
  Pre-existing Model-fitting Method
}

