\name{zelig2normal.gam}
\alias{zelig2normal.gam}
\title{Interface between the Zelig Model normal...}
\usage{zelig2normal.gam(formula, ..., data)}
\description{Interface between the Zelig Model normal.gam and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
