\name{zelig2poisson.gam}
\alias{zelig2poisson.gam}
\title{Interface between the Zelig Model poisson.gam and
the Pre-existing Model-fitting Method}
\usage{
  zelig2poisson.gam(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additional parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model poisson.gam and the
  Pre-existing Model-fitting Method
}

