\name{zelig2probit.gam}
\alias{zelig2probit.gam}
\title{Interface between the Zelig Model probit...}
\usage{zelig2probit.gam(formula, ..., data)}
\description{Interface between the Zelig Model probit.gam and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
