\docType{package}
\name{ZeligGAM-package}
\alias{ZeligGAM}
\alias{ZeligGAM-package}
\title{ZeligGAM}
\description{
  \tabular{ll}{ Package: \tab gam.zelig\cr Version: \tab
  0.2\cr Date: \tab 2011-04-25\cr Depends: Zelig (>=
  4.0-6), gam, mgcv\cr License: \tab GPL version 2 or
  newer\cr }
}
\details{
  Edit this description
}
\author{
  unknown
}
\keyword{package}

