\name{ZeligGAM-package}
\title{ZeligGAM...}
\description{ZeligGAM}
\details{\tabular{ll}{
Package: \tab gam.zelig\cr
Version: \tab 0.2\cr
Date: \tab 2011-04-25\cr
Depends: Zelig (>= 4.0-6), gam, mgcv\cr
License: \tab GPL version 2 or newer\cr
}

Edit this description}
\alias{ZeligGAM}
\docType{package}
\author{unknown}
\keyword{package}
